server/requirements.txt
bcrypt==3.1.4
passlib==1.7.1
psycopg2-binary==2.9.2
